+++ O24Sec/std_dict.pickle
